requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
nltk>=3.8.1
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=10.0.0
PyMuPDF>=1.25.0
python-multipart>=0.0.9
email-validator>=2.0.0
psycopg2-binary>=2.9.9
beautifulsoup4>=4.12.0
